cublas mvm. Spectrum as floats
adds ions as integers as usual, converts them to float after.
Not using floats straight off because atomicMax only works for integers.
A different method of adding ions may change this restriction.

memory wise, size of float is same as integer anyway.

+++ src/cuda/ion_series/check_series.cu
@@ -107,7 +107,7 @@ genMIons
  */
 void
 getSpecNonParallel(
-    uint32_t            *d_out_check_spec, 
+    float               *d_out_check_spec, 
     const uint8_t       *d_in_mions, 
     const float         *d_residual_raw, 
     const float         *d_mass_raw,
